Print Summary Popup 2 digit precision

+++ TFT/src/User/API/Printing.c
@@ -251,7 +251,7 @@ void endPrinting(void)
   u8  hour = infoPrinting.time/3600,
       min = infoPrinting.time%3600/60,
       sec = infoPrinting.time%60;
-  u16 meter = ((infoFile.source == BOARD_SD) ? coordinateGetAxisActual(E_AXIS) : coordinateGetAxisTarget(E_AXIS))/1000;
+  float meter = ((infoFile.source == BOARD_SD) ? coordinateGetAxisActual(E_AXIS) : coordinateGetAxisTarget(E_AXIS))/1000;
   sprintf(tempstr1, (char *)textSelect(LABEL_PRINT_TOTAL_TIME), hour,min,sec);
   if (meter >0 )
   {
